fix: 기록 삭제 위치 변경
